requirements.txt
flask
pymongo
Flask-GraphQL
graphene-mongo
MongoEngine
mongomock
flask-cors